Add python-based plugin-architecture (aka complete rewrite)
Turns out, the original idea to use MySQL views to merge
sample data from different weather station (databases) was
a terrible design decision because:

- All sample data must be stored in MySQL
- Tables must exist in the same database
- One-SQL-query-to-rule-them-all usually ends up a huge mess
- MySQL eats a lot of RAM, this is step one of migrating
  all weather station databases to sqlite or mangodb

This "rewrite" reworks this scheme to use a python class for
each type of weather station. These are referred to as "sources"
in th code. Under the hood, each source may use mysql or other
means of getting the data.

Winddb only handles the merging of data, outputting it to JSON
and running any callbacks.

This rewrite also drops support for XML. It was never a good idea
for winddb anyway, and the implementation was horrible. Probably
should have been done on a separate branch, but the diff was too
messy to work out once the crap-removal started and I'm lazy.