Separate sample type output and add ViVa-source support
This major patch introduces a new output format where samples of different
type are no longer grouped together by sample time.
This change was necessary to support sources where different types of samples
are not sampled at the same time or frequency.
The new output format is not backwards-compatible with the old format and
users of the API must be adapted.

This patch also adds support for ViVa weather stations.